package/libevas-generic-loaders: move to package directory
As for expedite package, there is no advantage for efl related
packages to share the same version number anymore
(except for efl and libelementary).

Also, we usually do not have a sub-directory for a family of related
packages which doen't share the same version number, so move
libevas-generic-loaders to package directory.
Libevas-generic-loaders appear now in "Libraries" -> "Graphics" in
the Kconfig menu.

Since libevas-generic-loaders now select the efl package, propagate
the reverse dependencies on mmu and wchar.

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
---
v4: improve the commit log (ThomasP)
    move libevas-generic-loaders to "Libraries" -> "Graphics" (ThomasP)
    fix package comment dependency on BR2_USE_MMU