package/efl/libefl: add webp dependency
WebP is a new image format that provides lossless and lossy
compression for images on the web. So enabling webp support
in efl libraries allow to loads images using WebP.

Also, it one of the "highly recommended" dependecies [1]
according to the README but disabling it doesn't need
the --enable-i-really-know-what-i-am-doing... option.
That's why BR2_PACKAGE_LIBEFL_WEBP is not added to
BR2_PACKAGE_LIBEFL_RECOMMENDED_CONFIG.

[1] https://git.enlightenment.org/core/efl.git/tree/README?id=v1.15.2#n486

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
---
v4: rename to libefl
    improve commit log