linux/perf: add optional dependency on newt
Remove slang dependency since the newt package depends on slang.
Perf build system disable automatically slang support if newt is not
installed.

Note: Don't trust the "Auto-detecting system features" tab, it only
detect that slang is installed on the target system but the slang
support is silently disabled if newt is not found.

Auto-detecting system features:
...                         dwarf: [ on  ]
...                         glibc: [ on  ]
...                          gtk2: [ OFF ]
...                      libaudit: [ OFF ]
...                        libbfd: [ OFF ]
...                        libelf: [ on  ]
...                       libnuma: [ on  ]
...                       libperl: [ OFF ]
...                     libpython: [ OFF ]
...                      libslang: [ on  ]
...                     libunwind: [ on  ]
...            libdw-dwarf-unwind: [ on  ]
...                          zlib: [ OFF ]
...                          lzma: [ OFF ]

Signed-off-by: Romain Naour <romain.naour@openwide.fr>