package/efl: rename libefl to efl
In order to avoid duplicate dependencies between efl and
libefl [1] change the BR2_PACKAGE_EFL from a menuconfig
to a simple menu.

This allow to rename the libefl package to efl to keep
backward compatibility with older defconfig files.

Add efl dependency on libelementary package.

Don't add Config.in.legacy entry for libefl since it's
not part of any Buildroot release.

[1] http://lists.busybox.net/pipermail/buildroot/2015-December/147128.html

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Cc: "Yann E. MORIN" <yann.morin.1998@free.fr>