package/skeleton: split into sysv and custom skeleton
Currently, both the sysv-like and custom skeletons are managed in a
single package.

However, for the custom skeleton, we basically do nothing, except for
a few symlinks for the /lib directory aliases, and a few checks about
merged /usr.

OTOH, for the default (sysv-like) skeleton, we do a lot more things,
like tweaking the issue, the hostname, setting the root password and so
on...

Furthermore, we're soon gonna have a skeleton for each kind of init
system, be it sysv-based or systemd.

So, it makes sense to offload the custom skeleton management to its own
package.

We change the way we check for the merged /usr readiness, from checking
at make parse time, to a configure command. This has virtually zero
impact, since the skeleton is the very first package to be "built", so
its configure commands are the very first to be run. It however slightly
simplifies the .mk as we need not enclose the parse-time checks inside
an ifeq-block to check if the package is enabled.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Romain Naour <romain.naour@openwide.fr>

---
Changes v2 -> v3:
  - fix install target macro  (Romain)
  - typoes  (Romain)