package/skeleton: make it a virtual package
The skeleton package now only serves as an intermediate package to
direct to the real skeleton implementation.

This is exactly what a virtual package is. Make it so.

Note: when graphing the dependency chain, and contrary to other virtual
packages, skeleton is de facto excluded from the dependencies of all
packages, and is "attached" as a dependency of the "All" package
instead, so as not to make the graph overly complicated and unreadable.

So, for it to appear on the graph, it needs to have a
BR2_PACKAGE_SKELETON kconfig variable, so that it appears in the output
of "make show-targets", like we do for the toolchain.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>

---
Changes v2 -> v3:
  - fix dependency  (Romain)
  - typoes  (Romain)