br2-external: source external-defconfig
As reported by Patrick Williams, there is not way to source an external
Makefile from BR2_EXTERNAL when BR2_HAVE_DOT_CONFIG is not set.

This is a problem for the support of merged defconfigs which should be
handled outside of Buildroot.

Add a new required file in BR2_EXTERNAL named external-defconfig.mk.
If BR2_EXTERNAL is not set, source a dummy external-defconfig.mk.

[1] http://lists.busybox.net/pipermail/buildroot/2016-July/166955.html
[2] http://patchwork.ozlabs.org/patch/641838/

Signed-off-by: Romain Naour <romain.naour@gmail.com>
---
TODO: add some Documentation