package/elf2flt: bump version to fix build with Binutils >= 2.32
Binutils added bfd_stdint.h to bfd.h [1], so elf2flt must create a
symlink for this header.

Bump the elf2flt version that include the upstream fix [2].

Keep our local patch 0001-ld-elf2flt-behave-properly-when-called-with-a-name-d.patch
that was not integrated with the recent project activity [3].

Fixes:
[armv7m-uclibc]
https://gitlab.com/kubu93/toolchains-builder/-/jobs/148356410

[m68k-coldfire-uclibc]
https://gitlab.com/kubu93/toolchains-builder/-/jobs/148356412

[1] https://sourceware.org/git/gitweb.cgi?p=binutils-gdb.git;a=commitdiff;h=2d5d5a8f0a8b5a03454bf168b7fa7024bb1ebbd8
[2] https://github.com/uclinux-dev/elf2flt/commit/429521f695c6f4212d809a141bfa4417b733f395
[3] https://github.com/uclinux-dev/elf2flt/pull/8

Signed-off-by: Romain Naour <romain.naour@gmail.com>
---
v2: bump the version instead of adding a new local patch (Arnout).